# admin/api/internal/openapi/swagger.json
{
    "components": {
        "schemas": {
...
            "http.ErrorBody": {
                "properties": {
                    "error": {
                        "description": "Error is a stable, machine-readable code.",
                        "type": "string"
                    },
                    "message": {
                        "description": "Message is a human-readable explanation. It never carries internal detail.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kube.Event": {
                "properties": {
                    "count": {
                        "type": "integer"
                    },
                    "lastSeen": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "object": {
                        "type": "string"
                    },
                    "reason": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kube.Filesystem": {
                "description": "Filesystem is the node's root disk. Nil unless node stats are switched on:\nreading them needs a grant on the kubelet that the panel does not hold by\ndefault. See charts/admin/templates/api/rbac.yaml.",
                "properties": {
                    "availableBytes": {
                        "type": "integer"
                    },
                    "capacityBytes": {
                        "type": "integer"
                    },
                    "usedBytes": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "kube.Namespace": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kube.Node": {
                "properties": {
                    "allocatable": {
                        "$ref": "#/components/schemas/kube.Resources"
                    },
                    "capacity": {
                        "$ref": "#/components/schemas/kube.Resources"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "filesystem": {
                        "$ref": "#/components/schemas/kube.Filesystem"
                    },
                    "name": {
                        "type": "string"
                    },
                    "os": {
                        "type": "string"
                    },
                    "pressure": {
                        "description": "Pressure lists the conditions that are true right now and should not be —\nMemoryPressure, DiskPressure, PIDPressure. Empty is the healthy case.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "ready": {
                        "type": "boolean"
                    },
                    "requested": {
                        "$ref": "#/components/schemas/kube.Resources"
                    },
                    "roles": {
                        "description": "Roles come from the node-role.kubernetes.io/* labels, which is the only\nplace Kubernetes records them.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "status": {
                        "description": "Status is the summary an operator reads first: Ready, NotReady, or\nReady,SchedulingDisabled for a cordoned node, which the conditions alone do\nnot say.",
                        "type": "string"
                    },
                    "usage": {
                        "$ref": "#/components/schemas/kube.Resources"
                    },
                    "version": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "kube.Pod": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "node": {
                        "type": "string"
                    },
                    "phase": {
                        "description": "Phase is Kubernetes' own phase, and Status is the more useful summary —\n\"Terminating\" or a waiting container's reason, which the phase does not say.",
                        "type": "string"
                    },
                    "ready": {
                        "description": "Ready counts containers, not pods: \"1/2\" is a running pod that is not\nserving, which is the case worth seeing.",
                        "type": "string"
                    },
                    "restarts": {
                        "type": "integer"
                    },
                    "status": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "kube.Resources": {
                "description": "Usage is the sum over the nodes that reported one, and is meaningless\nunless Summary.MetricsAvailable is true — a sum of nothing is zero, and\nzero is also what a genuinely idle cluster reports. This is a plain value\nrather than the pointer Node.Usage is because a total has one flag covering\nit, where a per-node reading has to say which nodes were measured.\n\nEven with MetricsAvailable true, a node that metrics-server has not sampled\nyet contributes nothing, so this can understate a cluster that is still\nwarming up.",
                "properties": {
                    "cpuMillis": {
                        "type": "integer"
                    },
                    "ephemeralBytes": {
                        "type": "integer"
                    },
                    "memoryBytes": {
                        "type": "integer"
                    },
                    "pods": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "kube.Storage": {
                "properties": {
                    "claims": {
                        "items": {
                            "$ref": "#/components/schemas/kube.VolumeClaim"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "volumes": {
                        "items": {
                            "$ref": "#/components/schemas/kube.Volume"
                        },
                        "type": "array",
                        "uniqueItems": false
                    }
                },
                "type": "object"
            },
...
            "kube.Summary": {
                "properties": {
                    "metricsAvailable": {
                        "description": "MetricsAvailable reports whether metrics-server answered. When false the\nusage figures are zero because nothing measured them, not because nothing\nis running — a distinction the dashboard has to draw for the operator.",
                        "type": "boolean"
                    },
                    "namespaces": {
                        "type": "integer"
                    },
                    "nodes": {
                        "$ref": "#/components/schemas/kube.NodeSummary"
                    },
                    "pods": {
                        "$ref": "#/components/schemas/kube.PodSummary"
                    },
                    "storage": {
                        "$ref": "#/components/schemas/kube.StorageSummary"
                    },
                    "workloads": {
                        "$ref": "#/components/schemas/kube.WorkloadSummary"
                    }
                },
                "type": "object"
            },
...
            "kube.VolumeClaim": {
                "properties": {
                    "accessModes": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "capacityBytes": {
                        "type": "integer"
                    },
                    "createdAt": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "requestedBytes": {
                        "description": "Requested is what the claim asked for and Capacity what it actually got.\nA provisioner may round up, so the two disagreeing is normal rather than a\nfault — but only Capacity says how much space there really is.",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Status is the claim's phase: Bound, Pending, or Lost. A Pending claim is\nusually the reason a StatefulSet's pod will not start.",
                        "type": "string"
                    },
                    "storageClass": {
                        "type": "string"
                    },
                    "volumeName": {
                        "description": "VolumeName is the PersistentVolume bound to this claim, empty while pending.",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "kube.Workload": {
                "properties": {
                    "createdAt": {
                        "type": "string"
                    },
                    "desired": {
                        "description": "Desired is how many replicas the controller wants, and Ready how many are\nserving. A DaemonSet's desired count is per node rather than configured.",
                        "type": "integer"
                    },
                    "images": {
                        "description": "Images are the container images the pod template runs, which is what says\nwhich version is deployed.",
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "uniqueItems": false
                    },
                    "kind": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "namespace": {
                        "type": "string"
                    },
                    "ready": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "An OAuth 2.1 access token from the configured provider.",
                "scheme": "bearer",
                "type": "http"
            }
        }
    },
    "info": {
        "description": "Manages the home lab's PostgreSQL and MongoDB services, and reads its Kubernetes cluster.\nEvery /api endpoint needs a bearer access token from the configured OpenID Connect\nprovider. There are no API keys: a caller acts as the person who signed in, and an\nagent carries that person's token.",
        "license": {
            "name": "MIT"
        },
        "title": "home-lab Admin API",
        "version": "0.1.0"
    },
...
    "paths": {
        "/api/kubernetes/namespaces": {
            "get": {
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/kube.Namespace"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List namespaces",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/namespaces/{namespace}/events": {
            "get": {
                "description": "Most recent first, capped at 100. Events carry the reason a pod is stuck\nwhen the pod itself only reports the symptom.",
                "parameters": [
                    {
                        "description": "Namespace",
                        "in": "path",
                        "name": "namespace",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/kube.Event"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List events",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/namespaces/{namespace}/pods": {
            "get": {
                "description": "Reports a summary alongside the phase. The phase alone does not say that a\nRunning pod is serving nothing, or that one is being deleted.",
                "parameters": [
                    {
                        "description": "Namespace",
                        "in": "path",
                        "name": "namespace",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/kube.Pod"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List pods",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/namespaces/{namespace}/workloads": {
            "get": {
                "description": "Deployments, StatefulSets, and DaemonSets as one list. The question is\nwhat is running, not which kind of controller runs it.",
                "parameters": [
                    {
                        "description": "Namespace",
                        "in": "path",
                        "name": "namespace",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/kube.Workload"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List workloads",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/nodes": {
            "get": {
                "description": "Capacity, what pods have reserved against it, and — when metrics-server\nis installed — what is actually being used. A node whose usage is absent\nreports no reading rather than zero.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/kube.Node"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "List nodes",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/storage": {
            "get": {
                "description": "Persistent volume claims and the volumes behind them. Both, because a\nreleased volume still holding data appears in neither list alone.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kube.Storage"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Read storage",
                "tags": [
                    "kubernetes"
                ]
            }
        },
        "/api/kubernetes/summary": {
            "get": {
                "description": "Counts and totals across nodes, pods, workloads, and storage, in one call\nso the dashboard neither renders nor fails in pieces.\n\nnodes.usage is a sum over the nodes that reported a reading, so it is zero\nboth for an idle cluster and for one nothing measured. metricsAvailable is\nwhat tells those apart: when it is false, treat nodes.usage as absent\nrather than as a measurement of zero.",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/kube.Summary"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/http.ErrorBody"
                                }
                            }
                        },
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Read cluster summary",
                "tags": [
                    "kubernetes"
                ]
            }
        },
...
    },
    "openapi": "3.1.0"
}